Feature/long text display optimization

See merge request enzoevers/ledmatrixdriver!4

+++ Code/LedMatrixBusLargeDriver/src/ledMatrixDriver/targetDisplays/ILedMatrix.h
@@ -5,8 +5,8 @@
 class ILedMatrix {
   public:
     virtual ~ILedMatrix() {};
-    virtual uint8_t getWidth() = 0;
-    virtual uint8_t getHeight() = 0;
+    virtual uint16_t getWidth() = 0;
+    virtual uint16_t getHeight() = 0;
     virtual void clearDisplay() = 0;
     virtual void fillDisplay() = 0;
     virtual void enableDisplay(bool state) = 0;
